+++ backend/canisters/user/api/src/updates/c2c_send_messages.rs
@@ -1,6 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{ChatId, EventIndex, MessageContent, MessageId, MessageIndex, MultiUserChat};
+use types::{Chat, ChatId, EventIndex, MessageContent, MessageId, MessageIndex};
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct Args {
@@ -27,7 +27,7 @@ pub enum Response {
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub enum C2CReplyContext {
     ThisChat(MessageId),
-    OtherChat(MultiUserChat, Option<MessageIndex>, EventIndex),
+    OtherChat(Chat, Option<MessageIndex>, EventIndex),
     OtherEventList(ChatId, Option<MessageIndex>, EventIndex),
 }