+++ backend/canisters/user/api/src/updates/mod.rs
@@ -48,6 +48,7 @@ pub mod set_avatar;
 pub mod set_bio;
 pub mod set_contact;
 pub mod set_message_reminder;
+pub mod set_message_reminder_v2;
 pub mod unarchive_chat;
 pub mod unblock_user;
 pub mod undelete_messages;