+++ frontend/openchat-agent/src/services/groupIndex/candid/types.d.ts
@@ -1062,7 +1062,7 @@ export type RemoveHotGroupExclusionResponse = { 'ChatNotFound' : null } |
   { 'Success' : null } |
   { 'InternalError' : string };
 export interface ReplyContext {
-  'chat_if_other' : [] | [[MultiUserChat, [] | [MessageIndex]]],
+  'chat_if_other' : [] | [[Chat, [] | [MessageIndex]]],
   'event_list_if_other' : [] | [[ChatId, [] | [MessageIndex]]],
   'event_index' : EventIndex,
 }