+++ frontend/openchat-agent/src/services/storageIndex/candid/types.d.ts
@@ -1000,7 +1000,7 @@ export type RemoveAccessorResponse = { 'Success' : null };
 export interface RemoveUserArgs { 'user_id' : UserId }
 export type RemoveUserResponse = { 'Success' : null };
 export interface ReplyContext {
-  'chat_if_other' : [] | [[MultiUserChat, [] | [MessageIndex]]],
+  'chat_if_other' : [] | [[Chat, [] | [MessageIndex]]],
   'event_list_if_other' : [] | [[ChatId, [] | [MessageIndex]]],
   'event_index' : EventIndex,
 }