+++ frontend/openchat-agent/src/services/userIndex/candid/types.d.ts
@@ -1071,7 +1071,7 @@ export type RemovePlatformModeratorResponse = { 'Success' : null } |
 export interface RemovePlatformOperatorArgs { 'user_id' : UserId }
 export type RemovePlatformOperatorResponse = { 'Success' : null };
 export interface ReplyContext {
-  'chat_if_other' : [] | [[MultiUserChat, [] | [MessageIndex]]],
+  'chat_if_other' : [] | [[Chat, [] | [MessageIndex]]],
   'event_list_if_other' : [] | [[ChatId, [] | [MessageIndex]]],
   'event_index' : EventIndex,
 }